requirements.txt
torch~=2.7.0+cu128
numpy~=2.2.2
matplotlib~=3.10.0
PyYAML~=6.0.2
tqdm~=4.67.1
scipy~=1.15.0